server/requirements.txt
# FastAPI 服务端依赖
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# 数据库
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Redis
redis[hiredis]==5.1.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP 客户端（调用微信 API / AI Provider）
httpx==0.27.2

# AI SDK（统一走 OpenAI 兼容协议，DeepSeek/通义都兼容）
openai==1.51.0

# 日志
loguru==0.7.2

# 工具
python-dotenv==1.0.1
orjson==3.10.7

# 测试
pytest==8.3.3
pytest-asyncio==0.24.0
